requirement.txt
# Web framework
Flask==3.1.1
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
blinker==1.9.0
colorama==0.4.6

# Data processing & images
numpy==2.3.1
pillow==11.3.0
opencv-python==4.8.0

# Deep learning
torch==2.1.0
torchvision==0.16.1
facenet-pytorch==2.8.2
faiss-cpu==1.7.4
